Add NetworkX BFS graph visualization script
This commit introduces a new Python script for visualizing graphs using the breadth-first search algorithm and NetworkX library. It includes functions to perform BFS on a given map and to visualize a graph. An example of its usage with two different graphs is also provided at the end of the script.